requirements.txt
streamlit==1.24.0
openai==1.3.5

pdfplumber==0.8.0
python-docx==0.8.11

numpy>=1.26.4
scikit-learn==1.3.0

tqdm==4.65.0
python-dotenv==1.0.0

pandas==2.2.0
pytest==7.4.0
